+++ src/ScottPlot4/ScottPlot/Renderable/AxisDimensions.cs
@@ -229,6 +229,9 @@ namespace ScottPlot.Renderable
             if (LockedLimits)
                 return;
 
+            if (min >= max)
+                throw new InvalidOperationException("Axis alimis: Max must be greater than Min");
+
             HasBeenSet = true;
             Min = min ?? Min;
             Max = max ?? Max;