+++ src/ScottPlot5/ScottPlot5/DataSources/IScatterSource.cs
@@ -1,4 +1,4 @@
-﻿namespace ScottPlot.DataSource;
+﻿namespace ScottPlot.DataSources;
 
 /// <summary>
 /// This interface is used by plottables to access data while rendering.