+++ src/ScottPlot5/ScottPlot5/DataSources/ScatterSourceXsYs.cs
@@ -1,4 +1,4 @@
-﻿namespace ScottPlot.DataSource;
+﻿namespace ScottPlot.DataSources;
 
 /// <summary>
 /// This data source manages X/Y points as separate X and Y collections